Use long instead of int as the return value of the NativeCrypto.SSL_do_handshake() to fix a crash in 64-bit builds which uses SSL. (fixes #739)

+++ rt/libcore/crypto/src/main/java/org/conscrypt/NativeCrypto.java
@@ -973,7 +973,7 @@ public final class NativeCrypto {
      * a server negotiation, supplying the {@code alpnProtocols} will enable
      * ALPN negotiation.
      */
-    public static native int SSL_do_handshake(long sslNativePointer,
+    public static native long SSL_do_handshake(long sslNativePointer,
                                               FileDescriptor fd,
                                               SSLHandshakeCallbacks shc,
                                               int timeoutMillis,
